skeleton script directory for tests and bw plotting
