+++ .gitignore
@@ -1,5 +1,6 @@
 # Custom
 *~
+*__pycache__
 build*/
 debug*/
 utils/
