+++ scripts/plot_test_bw.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+def plot_test_bw(in_file, out_file):
+    print(in_file, out_file)
